+++ R/ha_coverage.R
@@ -20,43 +20,32 @@
 #' ha_coverage("POP", progress = FALSE)
 #' }
 ha_coverage <- function(topic_key, layer_key = NULL, keys_only = FALSE, progress = TRUE) {
+  chk::chk_not_missing(topic_key, x_name = "`topic_key`")
+  chk::chk_string(topic_key)
+  if (!is.null(layer_key)) {
+    chk::chk_character(layer_key)
+  }
+  chk::chk_logical(keys_only)
+  chk::chk_logical(progress)
+
   body <- ha_api_coverage_req(topic_key, layer_key) |>
     ha_req_perform() |>
     ha_resp_body("coverages")
 
-  output <- tibble::tibble(body) |>
-    tidyr::unnest_longer(body) |>
-    tidyr::unnest_longer(body) |>
-    tidyr::unnest_wider(body) |>
-    dplyr::mutate("topic_key" = topic_key) |>
-    dplyr::select(
-      c(
-        "topic_key",
-        "population_key" = "population",
-        "period_key" = "period",
-        "layer_key" = "body_id"
-      )
-    )
+  output <- cbind(topic_key, body)
+  colnames(output) <- c("topic_key", "population_key", "period_key", "layer_key")
 
   if (!keys_only) {
     stratifications <- ha_stratifications(progress)
     layers <- ha_layers()
 
     output <- output |>
-      dplyr::right_join(stratifications, by = "population_key") |>
-      dplyr::right_join(layers, by = "layer_key") |>
-      dplyr::select(
-        c(
-          "topic_key",
-          "population_key",
-          "population_name",
-          "population_grouping",
-          "period_key",
-          "layer_key",
-          "layer_name"
-        )
-      )
+      merge(stratifications, by = "population_key", all.x = TRUE) |>
+      merge(layers, by = "layer_key", all.x = TRUE)
+      
+    output <- output[c("topic_key", "population_key", "population_name", 
+      "population_grouping", "period_key", "layer_key", "layer_name")]
   }
 
-  output
+  tibble::as_tibble(output)
 }