+++ R/helpers.R
@@ -1,9 +1,22 @@
-ha_req <- function(endpoint) {
-  if (!curl::has_internet()) {
-    stop("Your API call has errors. No Internet Connection.")
-  } else if (is.null(ha_get())) {
-    stop("Set your portal using ha_set()")
+chk_has_internet <- function() {
+  if (vld_has_internet()) {
+    return(invisible())
   }
+  chk::abort_chk("Your API call has errors. No Internet Connection.")
+}
+vld_has_internet <- function() curl::has_internet()
+
+chk_has_set_ha <- function() {
+  if (vld_has_set_ha()) {
+    return(invisible())
+  }
+  chk::abort_chk("Health atlas portal must be set. Try using ha_set()")
+}
+vld_has_set_ha <- function() !is.na(ha_get())
+
+ha_req <- function(endpoint) {
+  chk_has_internet()
+  chk_has_set_ha()
 
   httr2::request(ha_get()) |>
     httr2::req_user_agent("healthatlas R package") |>
@@ -20,11 +33,10 @@ ha_req_perform_iterative <- function(req, progress = TRUE) {
       start = 0,
       offset = 20,
       resp_pages = function(resp) {
-        count <- httr2::resp_body_json(resp) |>
-          purrr::pluck("count")
+        count <- httr2::resp_body_json(resp)[["count"]]
 
         if (count == 0) {
-          stop("Your API call has errors. No Results.")
+          chk::abort_chk("Your API call has errors. No Results.")
         }
         ceiling(count / 20)
       }
@@ -40,23 +52,28 @@ ha_req_perform <- function(req) {
 
 ha_resp_body_iterative <- function(resp) {
   resp |>
-    purrr::map(httr2::resp_body_json) |>
-    purrr::map(~ .x$results) |>
-    purrr::list_c()
+    lapply(\(x) httr2::resp_body_json(x, simplifyVector = TRUE)) |>
+    lapply(\(x) x[["results"]]) |>
+    do.call(what = rbind)
 }
 
-ha_resp_body <- function(resp, accessor = NULL) {
-  body <- httr2::resp_body_json(resp)
+ha_resp_body <- function(resp, accessor) {
+  body <- httr2::resp_body_json(resp, simplifyVector = TRUE)
 
-  count <- purrr::pluck(body, "count")
+  count <- body[["count"]]
   if (!is.null(count) && count == 0) {
-    stop("Your API call has errors. No Results.")
+    chk::abort_chk("Your API call has errors. No Results.")
   }
-  if (!is.null(accessor)) {
-    body <- purrr::pluck(body, accessor)
+  body <- body[[accessor]]
+  if (is.matrix(body)) {
+    body <- as.data.frame(body)
   }
-  if (!is.null(names(body))) {
-    body <- list(body)
+  if (!is.data.frame(body)) {
+    body <- body[lengths(body) != 0]
+    body <- lapply(body, as.data.frame)
+    body <- lapply(names(body), \(x) cbind(body[[x]], x))
+    body <- do.call(rbind, body)
   }
+
   body
 }