+++ inst/doc/typical_usage.R
@@ -38,20 +38,51 @@ ease_of_access <- ha_data(
 ease_of_access
 
 ## -----------------------------------------------------------------------------
+combinations_of_data <- ha_data(
+  topic_key = c("POP", "UMP"),
+  population_key = c("", "H"),
+  period_key = c("2017-2021", "2018-2022", "invalid"),
+  layer_key = "neighborhood"
+)
+combinations_of_data
+
+## -----------------------------------------------------------------------------
+library(tibble)
+library(purrr)
+
+# creating a table of data I want
+metadata <- tribble(
+  ~ topic_key, ~ population_key, ~ period_key, ~ layer_key,
+  "POP",       "",               "2017-2021",  "neighborhood",
+  "HCSFVAP",   "",               "2020-2021",  "neighborhood",
+  "UMP",       "H",              "2017-2021",  "neighborhood",
+)
+
+metadata %>%
+  pmap(ha_data)
+
+## -----------------------------------------------------------------------------
 layers <- ha_layers()
 layers
 
 ## -----------------------------------------------------------------------------
 community_areas <- ha_layer("neighborhood")
+community_areas
 
 ## -----------------------------------------------------------------------------
-library(dplyr)
+ease_of_access <- ha_data(
+  topic_key = "HCSFVAP",
+  population_key = "",
+  period_key = "2022-2023",
+  layer_key = "neighborhood",
+  geometry = TRUE
+)
+ease_of_access
+
+## -----------------------------------------------------------------------------
 library(ggplot2)
 
-map_data <- community_areas |>
-  left_join(ease_of_access, "geoid") 
-
-plot <- ggplot(map_data) +
+plot <- ggplot(ease_of_access) +
   geom_sf(aes(fill = value), alpha = 0.7) +
   scale_fill_distiller(palette = "GnBu", direction = 1) +
   labs(