+++ NEWS.md
@@ -1,5 +1,10 @@
-# healthatlas 0.1.0
+# healthatlas 0.1.1
 
-# bcdata 0.1.0
+* `dplyr`, `tidyr`, and `purrr` dependencies were removed (#2).  
+* `chk`s were introduced to ensure type checking and provide consistent error
+syntax.  
+* `ha_get()` now returns `NA` if no health atlas is set.
+
+# healthatlas 0.1.0
 
 * Initial Release